Extract getDropSnappedTime function to useTimelinePositioning hook
- Moved getDropSnappedTime function from timeline-track.tsx (lines 376-420)
- Added useTimelinePositioning hook with all necessary imports
- Updated timeline-track.tsx to use the extracted function
- No TypeScript errors, ready for testing
